Add a basic template for case studies
