+++ app/templates/ai_lab/ai_lab_case_study.html
@@ -1,0 +1,31 @@
+{% extends "base.html" %}
+{% load wagtailcore_tags static wagtailuserbar nhsukfrontend_tags %}
+
+{% block body_class %}template-article{% endblock %}
+
+{% block extra_css %}
+{% endblock extra_css %}
+
+{% block content %}
+{% breadcrumb %}
+<div class="nhsuk-width-container">
+  <main class="nhsuk-main-wrapper" id="maincontent">
+    <div class="nhsuk-grid-row">
+      <div class="nhsuk-grid-column-two-thirds">
+        <h1 class="nhsuk-heading-xl nhsuk-u-margin-bottom-5">{{ page.title }}</h1>
+        {% if show_meta %}
+        {% include '_partials/page_meta.html' %}
+        {% endif %}
+      </div>
+    </div>
+    <div class="nhsuk-grid-row">
+      <div class="nhsuk-grid-column-full">
+        {% include_block page.body %}
+
+        {% block extra_content %}
+        {% endblock %}
+      </div>
+    </div>
+  </main>
+</div>
+{% endblock content %}
